// app/(auth)/shared/types/enums.ts
export enum UserEnum {
  MENTEE = 'MENTEE',
  MENTOR = 'MENTOR',
  EXPERIENCED = 'EXPERIENCED',
  NEWBIE = 'NEWBIE',
}

export enum SocialMedia {
  GOOGLE = 'GOOGLE',
  GITHUB = 'GITHUB',
  LINKEDIN = 'LINKEDIN',
}

export enum InputEnum {
  TEXT = 'TEXT',
  EMAIL = 'EMAIL',
  PASSWORD = 'PASSWORD',
  SELECT = 'SELECT',
  URL = 'URL',
  IMAGE = 'IMAGE',
  TEXTAREA = 'TEXTAREA',
}

export enum WaitList {
  TEXT = 'TEXT',
  EMAIL = 'EMAIL',
  SELECT = 'SELECT',
}